refactor(iam): align acceptance preflight boundary
